refactor: consolidate remote config service implementations
